Update generate-document with suggested feedback
The generateDocument function was placed into a seperate file
"chromium.ts" to mock the function and allow for a Buffer to be
returned during the unit test.

Ticket: AT-7345
