+++ document-generation/chromium.ts
@@ -1,0 +1,32 @@
+import chromium from "@sparticuz/chrome-aws-lambda";
+import { PDFOptions } from "puppeteer-core";
+import { logger } from "../utils/logging";
+
+export async function generateDocument(document: string): Promise<Buffer | undefined> {
+  let browser, generatedDocument;
+
+  try {
+    browser = await chromium.puppeteer.launch({
+      args: chromium.args,
+      defaultViewport: chromium.defaultViewport,
+      executablePath: await chromium.executablePath,
+      headless: chromium.headless,
+      ignoreHTTPSErrors: true,
+    });
+
+    const page = await browser.newPage();
+    const options: PDFOptions = { format: "A4" };
+
+    await page.setContent(document);
+    await page.emulateMediaType("screen");
+    generatedDocument = await page.pdf(options);
+    console.log("DOCU: ", "generateDocument");
+    logger.info("Document generation complete");
+  } catch (error) {
+    logger.error(error as any);
+  } finally {
+    browser?.close();
+  }
+
+  return generatedDocument;
+}
